bbai_mcp_client/pyproject.toml
[project]
name = "bbai-mcp-client"
version = "0.1.0"
description = "MCP Client for AI Content Publisher"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.9.0",
    "pymongo>=4.6.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "httpx>=0.25.0"
]
